feat: basic Solana wallet example (#1)
[(XC-281)](https://dfinity.atlassian.net/browse/XC-281) Basic wallet
example for SOL RPC demonstrating interaction with the Solana blockchain
and SPL tokens from a canister through the Solana JSON RPC interface.

The following functionalities are demonstrated in this example:
- Deriving Solana addresses for a principal
- Checking the balance of a Solana account
- Creating a nonce account and fetching the latest nonce from it
- Making a Solana transaction to transfer SOL between accounts using
either a recent blockhash or a durable nonce
- Creating an Associated Token Account (ATA) for a given SPL token,
fetching the account balance, and transferring SPL tokens between
accounts

Note that since the SOL RPC canister is not yet published, a basic
stand-in is provided that performs JSON RPC requests. This currently
only works for local development since it does not address the consensus
issues for obtaining a recent blockhash and uses the Solana Foundation
RPC provider which is not IPv6 compatible.

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>