feat: add logging crate (#13)
This PR adds the `canlog` and `canlog_derive` crates which wrap the
`ic_canister_log` crate to add support for log priorities, sorting, and
filtering. The main functionality is provided by the `canlog::log` macro
and `canlog::LogPriorityLevels` trait. The `canlog_derive` crate offers
a procedural macro to derive the `LogPriorityLevels` trait automatically
for an enum.