feat: round result from `getSlot` RPC method (#48)
(XC-292) Round the result from the Solana `getSlot` RPC method to the
nearest `n` blocks where `n` is a user parameter given as an input to
the canister method. This is done to partly circumvent the fast Solana
block time and achieve a higher consensus rate when performing the HTTP
outcalls to fetch the latest slot.

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>