feat: client builder (#54)
Refactor the `SolRpcClient` in the `sol_rpc_client` crate to use the
builder pattern for the client and the request to offer more
flexibility:

1. The general API follows that of
[`reqwest`](https://github.com/seanmonstar/reqwest).
2. There is builder for the client to configure options common to all
calls made to the SOL RPC canister.
3. There is a builder for each request created by the client, to allow
overriding any configuration (e.g. amount of cycles, consensus strategy,
etc) for that particular request.
4. Reserved a canister ID `tghme-zyaaa-aaaar-qarca-cai` for the future
productive SOL RPC canister to be able to instantiate a client targeting
that canister (most common use-case)