feat: cycles cost (#52)
Add endpoints to the SOL RPC canister to estimate the cycle cost of a
given request. The proposed API is as a follows:

1. A caller can pass the exact same request parameters to the cycles
cost estimation endpoint as to the endpoint that would actually execute
the request so that a precise cycles cost estimation can be returned for
that particular request.
2. Since the endpoints on the SOL RPC canister use different types for
the parameters, there is a dedicated `_CycleCost` endpoint for every
Solana endpoint supported by the canister.
3. The idea is that the caller could use those endpoints to establish a
baseline and then do the other calls for that endpoint with some
additional cycles to account for minor changes in request parameters
(unused cycles are reimbursed).

Implementation details:

1. Introduce the concept of `MultiRpcRequest`, which is a JSON-RPC
request aimed at multiple URLs (corresponding to Solana JSON-RPC
providers). This replaces the previous clients used in the canister
(`CandidRpcClient` and `SolRpcClient`).
2. This allows to implement once a `cycles_cost` method for all
`MultuRpcRequest` that we will support.