docs: Rust documentation tests for the `SolRpcClient` (#65)
Add Rust documentation tests for the `SolRpcClient` in the
`sol_rpc_client` crate.

This requires a mocked implementation of a `Runtime` to mock a canister
making (through the client) an update or a query call:

1. This implementation is only available behind the flag
`#[cfg(not(target_arch = "wasm32"))]` to ensure that it cannot be used
inside a canister and does not "pollute" the productive code.
2. An alternative solution would have been to extract this mocked
implementation to yet another crate; however, this was deemed to
cumbersome.