build: release pipeline (#4)
([XC-284](https://dfinity.atlassian.net/browse/XC-284)): Use
[release-plz](https://release-plz.dev/) to start automatizing the
release process:

1. Job can be manually triggered to create a new PR containing the new
versions and changelog for the next release. An example of such PR is
#83.
2. An opened PR should trigger the normal CI flow (in particular the
end-to-end tests with the deployment of the canister wasm built
reproducibly). However, a
[limitation](https://release-plz.dev/docs/github/token) of the default
`GITHUB_TOKEN` in actions is that it cannot trigger other workflows. An
easy workaround is to close/re-open the PR to trigger the necessary
jobs.

A future PR will use `release-plz release` to publish the required
crates on crates.io and create a Github release containing the wasm and
the Candid file.

[XC-284]:
https://dfinity.atlassian.net/browse/XC-284?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ