test: integration test for `verifyApiKey` (#82)
The endpoint `verifyApiKey` is useful for debugging and testing purposes
to ensure that an API key for a given provider is correctly set. This is
also obviously a sensitive endpoint and proper access control should be
in place to avoid security issues such as leaking the API key. An
integration test with Pocket IC was added to ensure that is the case.