feat: Add support for `getRecentPrioritizationFees` (#92)
Add support for the method
[`getRecentPrioritizationFees`](https://solana.com/de/docs/rpc/http/getrecentprioritizationfees).

Supporting this method via HTTPs outcalls offers similar challenges to
be able to achieve consensus on the responses as for supporting the
method [`getSlot`](https://solana.com/de/docs/rpc/http/getslot) (#33,
#48). Specifically,

1. Each response to `getRecentPrioritizationFees` consists of a
fast-changing subset of slots, where each slot is associated with a
[prioritization
fee](https://solana.com/de/developers/guides/advanced/how-to-use-priority-fees).
Responses between nodes may vary because the response contains
`processed` slots and a new `processed` slot is produced roughly every
400ms.
2. Similarly to `getSlot`, we need a heuristic to select
deterministically a subset of slots in the response seen by a node that
at least two thirds of nodes can agree on with a significant
probability, which proceeds as follows:
    1. Sort the prioritization fees according to their slot.
2. Agree on the last processed slot. This uses rounding, exactly as for
`getSlot`.
3. Agree on the size of the subset that will contain the prioritization
fee of the previously-agreed last processed slot and the remaining fees
by decreasing order of slot (higher is more recent).