feat: add support for `getSignatureStatuses` RPC method (#96)
(XC-291) Add support for `getSignatureStatuses` RPC method.

Due to the fast-changing value of the `confirmations` field in the
response of `getSignatureStatuses`, it is set to `null` in the response
transform. This was done to achieve consensus. Rounding is not performed
on this value since a value of `null` is anyways returned after a
supermajority of the cluster finalized the transaction, meaning its
value is short-lived and the rounded value would not be of much use.