feat: add client method to sign a transaction (#113)
(XC-317) Add a helper method in the `sol_rpc_client` module to signa a
Solana transaction using threshold EdDSA. This method is not part of the
client `struct` as it does not interact with the SOL RPC canister, but
instead with the management canister. However, it takes a runtime
argument to be re-usable in tests and prod code.