refactor: remove default/non-default providers (#122)
The notion of default/non-default providers was only useful when
`Equality` was chosen as consensus strategy and no providers was
explicitly specified, in which case the "default" providers would be
selected. This PR removes the notion of default/non-default providers by
replacing with an ordered list of providers for a given Solana cluster.
The previous notion of "default" providers simpy correspond to the first
three providers in that list.