chore: add optional `cost_units` to `TransactionStatusMeta` (#180)
In
[v2.3.1](https://docs.rs/solana-transaction-status/2.3.1/solana_transaction_status/struct.UiTransactionStatusMeta.html)
of the `solana-transaction-status-client-types` crate, the a new
optional `cost_units` field was added to `UiTransactionStatusMeta` which
is included in responses to the `getTransaction` RPC endpoint (see [this
PR](https://github.com/anza-xyz/agave/pull/5245)). Add a corresponding
field to `sol_rpc_types::TransactionStatusMeta` This should not be
breaking for the Candid API since the field is optional, and optional
fields are [ignored by
default](https://github.com/serde-rs/serde/pull/201) by `serde`.