build: release with multi-tags (#174)
Adapt the configuration of release-plz to use a tag/version per artifact
instead of having a single tag/version for the whole repository. This is
needed as the semantic versioning of libraries may change differently
(e.g. a breaking change for `sol_rpc_types` is not necessrily breaking
for `sol_rpc_client`). Also document how to make a release.