feat: Select supported providers based on successful responses (#183)
Dynamically select supported JSON-RPC providers according to the number
of successful JSON-RPC responses in the last 20 minutes, where providers
with a higher number will be prioritized over ones with lower numbers.

The reason for this heuristic is similar to dfinity/evm-rpc-canister#435
since JSON-RPC errors are also not part of the Solana RPC API and
therefore not standardized.