chore: do not record metrics for requests with insufficient cycles (#184)
(XC-416) Do not record metrics for requests with too few cycles. Namely:
- Never increment `solrpc_requests` or `solrpc_responses` metrics for
requests with too few cycles.
- Do not increment `solrpc_inconsistent_responses`metrics for requests
where all requests had too few cycles, since in this case it is expected
that the results are inconsistent due to the cycles cost of executing
JSON-RPC requests depending on the specific providers (due e.g. to the
varying target URL length and authentication methods).