ci: revert release-plz action to upstream repository (#221)
In #215 we switched to building the release-plz CLI from source to
include https://github.com/release-plz/release-plz/pull/2357 before it
was merged. Since that change is now merged in the upstream repository
master branch, this PR switches back to the upstream version of the
release-plz GitHub action, pinning the release-plz version to v0.3.142
which includes https://github.com/release-plz/release-plz/pull/2357.