chore: handle `max_response_bytes` is exceeded error in metrics (#235)
Currently, all `SYS_FATAL` IC error responses are labelled with error
type `ic` in the metrics. However, this error also occurs when the
response size `exceeds max_response_bytes`, which is a client error. As
such, a new `max-response-size-exceeded` error label for
`max_response_bytes` exceeded is added, and separated from other IC
errors.

This is a similar change to
[dfinity/evm-rpc-canister#463](https://github.com/dfinity/evm-rpc-canister/pull/463).